加入桌面小部件

+++ app/src/main/java/com/wm/remusic/service/MusicPlayer.java
@@ -692,7 +692,9 @@ public class MusicPlayer {
     }
 
     public static void clearQueue() {
+
         try {
+            if(mService != null)
             mService.removeTracks(0, Integer.MAX_VALUE);
         } catch (final RemoteException ignored) {
         }
